+++ tests/test_httpie.py
@@ -10,19 +10,19 @@ import httpie
 
 def test_debug():
     r = http('--debug')
-    assert r.exit_status == httpie.ExitStatus.OK
+    assert r.exit_status == httpie.ExitStatus.SUCCESS
     assert 'HTTPie %s' % httpie.__version__ in r.stderr
 
 
 def test_help():
     r = http('--help', error_exit_ok=True)
-    assert r.exit_status == httpie.ExitStatus.OK
+    assert r.exit_status == httpie.ExitStatus.SUCCESS
     assert 'https://github.com/jakubroztocil/httpie/issues' in r
 
 
 def test_version():
     r = http('--version', error_exit_ok=True)
-    assert r.exit_status == httpie.ExitStatus.OK
+    assert r.exit_status == httpie.ExitStatus.SUCCESS
     # FIXME: py3 has version in stdout, py2 in stderr
     assert httpie.__version__ == r.stderr.strip() + r.strip()